dom/tests/mochitest/pointerlock/file_userPref.html

<!DOCTYPE HTML>
<html>
  <!--
  https://bugzilla.mozilla.org/show_bug.cgi?id=633602
  -->
  <head>
    <title>Test for Bug 633602</title>
    <script type="application/javascript" src="/tests/SimpleTest/SimpleTest.js">
    </script>
    <script type="application/javascript" src="/tests/SimpleTest/EventUtils.js">
    </script>
    <script type="application/javascript" src="mouselock_util.js">
    </script>
    <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
  </head>
  <body onload="start();">
    <a target="_blank"
       href="https://bugzilla.mozilla.org/show_bug.cgi?id=633602">
      Mozilla Bug 633602</a>
    <p id="display"></p>
    <div id="content" style="display: none">
      <canvas id="canvas" width="150" height="150"></canvas>
    </div>
    <pre id="test">
      <script type="application/javascript">
        /*
         * Test for Bug 633602 
         * Test to ensure that pointer lock is allowed when the user preference
         * setting for pointer lock is true and the element is in fullscreen.
         */
        SimpleTest.waitForExplicitFinish();

        var pointer = navigator.mozPointer;
        var canvas = document.getElementById("canvas");

        document.addEventListener("mozfullscreenchange", function(e) {
          if (document.mozFullScreen &&
              document.mozFullScreenElement === canvas) {
            // Set Preference to True to see if lock succeeds and element is
            // in fullscreen
            SpecialPowers.setBoolPref("full-screen-api.pointer-lock.enabled",
                                      true );

            pointer.lock(canvas, function() {
              is(pointer.isLocked, true, "Mouse should be locked when " +
                 "pref setting is true and element is in fullscreen.");
              pointer.unlock();

              // Set Preference to False to see if lock fails and element is
              // in fullscreen
              SpecialPowers.setBoolPref("full-screen-api.pointer-lock.enabled",
                                        false);

              pointer.lock(canvas, function() {
                is(pointer.isLocked, true, "Success Callback - pointer should " +
                 "not be locked as user pref is set to false");
                SimpleTest.finish();
              }, function() {
                is(pointer.isLocked, false, "Mouse shouldn't be locked " +
                   "when pref setting is false and element is in fullscreen");
                document.mozCancelFullScreen();
                SpecialPowers.setBoolPref("full-screen-api.pointer-lock.enabled",
                                          true);
                SimpleTest.finish();
              });
            }, function() {
              is(pointer.isLocked, false, "Failure Callback - pointer should " +
                 "not be locked");
              SimpleTest.finish();
            });
          } else {
            SimpleTest.finish();
          }
        }, false);

        function start() {
          SimpleTest.waitForFocus(function() {
              canvas.mozRequestFullScreen();
          });
        }
      </script>
    </pre>
  </body>
</html>
